% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cases.R
\name{cases}
\alias{cases}
\title{Create an S7 object of class "cases"}
\usage{
cases(numerator = class_missing, denominator = class_missing)
}
\arguments{
\item{numerator}{A non-negative \link{numeric} vector of numbers of individuals.}

\item{denominator}{A non-negative \link{numeric} vector of numbers of individuals.}
}
\value{
An S7 object of class "cases".
}
\description{
A \code{cases} object contains numbers of individuals, typically
representing a public health outcome. This can for example
represent the number of Persons Who Inject Drugs (PWID) or the number
of people living with type 2 Diabetes.
}
\examples{
library(suppress)

cases(10) # 10 cases
cases(10, 100) # 10 cases out of a possible 100

numerator <- sample.int(10, 30, TRUE)
denominator <- rep(100, 30)

cases(numerator, denominator)

}
\seealso{
\link{suppress}
}
