% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suppress.R
\name{suppress}
\alias{suppress}
\title{Suppress sensitive numbers and percentages}
\usage{
suppress(x, ...)
}
\arguments{
\item{x}{A cases object.}

\item{...}{Additional arguments passed onto the 3 suppress functions:
\code{suppress_numerator}, \code{suppress_denominator}, and \code{suppress_percentage}.}
}
\value{
A data frame of suppressed numerator, denominator and percentage.
}
\description{
Suppress sensitive numbers and percentages
}
\details{
When only 1 value is suppressed, secondary suppression is applied to
the next smallest unsuppressed value.
}
\examples{
library(suppress)

# Create an S7 object of class sensitive
cases <- cases(0:10, rep(20, 11))

suppress(cases)
}
\seealso{
\link{suppress_numerator}
}
