% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suppress_utils.R
\name{numerator}
\alias{numerator}
\alias{denominator}
\alias{suppress_numerator}
\alias{suppress_denominator}
\alias{proportion}
\alias{percentage}
\alias{suppress_proportion}
\alias{suppress_percentage}
\alias{dummy}
\title{Suppress sensitive numbers and percentages}
\usage{
numerator(x, ...)

denominator(x, ...)

suppress_numerator(x, ...)

suppress_denominator(x, ...)

proportion(x, ...)

percentage(x, ...)

suppress_proportion(x, ...)

suppress_percentage(x, ...)

dummy(x, threshold = NULL, replace = NULL, digits = NULL)
}
\arguments{
\item{x}{A cases object.}

\item{...}{Additional arguments.}

\item{threshold}{A number such that any values less or equal to this are
regarded as sensitive and replaced with an appropriate value. \cr
The default is 5.
To set globally use \code{options(suppress.threshold)}}

\item{replace}{A value to replace numbers <= threshold. \cr
The default is \code{NA}. To set globally use \code{options(suppress.replace)}}

\item{digits}{Number of digits to round the percentages to. \cr
The default is 2. To set globally use \code{options(suppress.digits)}}
}
\value{
A vector the length of numerator or denominator, depending on which
method is called.
}
\description{
Suppress sensitive numbers and percentages
}
\details{
When only 1 value is suppressed, secondary suppression is applied to
the next smallest unsuppressed value. \cr

\code{dummy()} was created to temporarily satisfy the R CMD check. Don't use it.
}
\examples{
library(suppress)

# Create an S7 object of class sensitive
cases <- cases(0:10, rep(20, 11))

# Suppress numbers <= 5
suppress_numerator(cases)
suppress_numerator(cases, threshold = 7, replace = "*")
suppress_denominator(cases)

##### Reporting with  percentages #####

# If we reported denominator and percentages
# We could easily calculate the underlying numerators

p <- proportion(cases)

# A user could calculate the numerator this way
p * denominator(cases)

# Use suppress_percentage to
# ensure percentages are also suppressed
suppress_proportion(cases)
suppress_percentage(cases, replace = "*")

# Finally

secret_data <- data.frame(secret = numerator(cases),
                          denominator = denominator(cases),
                          percentage = percentage(cases))

published_data <- data.frame(secret = suppress_numerator(cases, replace = "*"),
                             denominator = suppress_denominator(cases),
                             percentage = suppress_percentage(cases, replace = "*"))
secret_data
published_data
}
\seealso{
\link{suppress}
}
